requirements.txt
GitPython==3.1.29
h5py==3.7.0
matlab==0.1
matlabengine==9.13.1
matplotlib==3.6.0
networkx==2.8.7
numpy==1.23.4
pandas==1.5.1
plotly==5.10.0
scipy==1.9.3
seaborn==0.12.1
